# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PGO Engine API",
    "description": "Secured with Basic Auth",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://pgo.otapp.live/api",
      "description": "Generated server url"
    }
  ],
  "security": [{ "basicAuth": [] }],
...
  "paths": {
...
    "/admin/v1/payment-channels/uid/{uid}": {
      "get": {
        "tags": ["Admin Payment Channels Management"],
        "summary": "Get payment channel by UID",
        "description": "Fetch a payment channel by its unique UID.",
        "operationId": "getPaymentChannelByUid",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment channel retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperPaymentChannelResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Admin Payment Channels Management"],
        "summary": "Update payment channel",
        "description": "Updates mutable details of a payment channel by UID. Channel code remains immutable and cannot be changed.",
        "operationId": "updatePaymentChannel",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentChannelRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Payment channel updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperPaymentChannelResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/merchants/uid/{uid}": {
      "get": {
        "tags": ["Admin Merchant Management"],
        "summary": "Get merchant by UID",
        "description": "Returns a merchant by its unique UID.",
        "operationId": "getMerchantByUid",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Merchant retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperMerchantResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Admin Merchant Management"],
        "summary": "Update merchant",
        "description": "Updates an existing merchant by UID.",
        "operationId": "updateMerchant",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMerchantRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Merchant updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperMerchantResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin Merchant Management"],
        "summary": "Delete merchant",
        "description": "Deletes a merchant by its database ID.",
        "operationId": "deleteMerchant",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Merchant deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/payment-gateways/uid/{gatewayUid}/channels/uid/{channelUid}": {
      "post": {
        "tags": ["Admin Payment Gateway Management"],
        "summary": "Assign payment channel to payment gateway",
        "description": "Links a payment channel to a payment gateway with payCode and provider metadata.",
        "operationId": "assignPaymentChannelToGateway",
        "parameters": [
          {
            "name": "gatewayUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "channelUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignChannelToGatewayRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Payment channel assigned to gateway successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperPaymentGatewayChannelResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/gateway-channels/gateway/uid/{gatewayUid}/channel/uid/{channelUid}": {
      "post": {
        "tags": ["Admin Gateway - Channel Management"],
        "summary": "Assign channel to gateway",
        "description": "Assigns a payment channel to a payment gateway with payCode and provider details.",
        "operationId": "assignChannelToGateway",
        "parameters": [
          {
            "name": "gatewayUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "channelUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignChannelToGatewayRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Channel assigned successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperPaymentGatewayChannelResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/users/uid/{userUid}": {
      "get": {
        "tags": ["Admin User Management"],
        "summary": "Get user by UID",
        "description": "Returns a user by their unique UID.",
        "operationId": "getUserByUid",
        "parameters": [
          {
            "name": "userUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperUserResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/transactions/uid/{uid}": {
      "get": {
        "tags": ["Admin Transaction Management"],
        "summary": "Get transaction by internal ID(Ready)",
        "description": "Returns a transaction by its internal transaction ID.",
        "operationId": "getTransactionByInternalId",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperTransactionResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/transactions/merchant/{merchantId}": {
      "get": {
        "tags": ["Admin Transaction Management"],
        "summary": "Get transaction by merchant ID",
        "description": "Returns a transaction by its merchant transaction ID.",
        "operationId": "getTransactionByMerchantId",
        "parameters": [
          {
            "name": "merchantId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperTransactionResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/payment-gateways/uid/{uid}": {
      "get": {
        "tags": ["Admin Payment Gateway Management"],
        "summary": "Get payment gateway by UID",
        "description": "Returns a payment gateway by its unique UID.",
        "operationId": "getPaymentGatewayByUid",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment gateway retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperPaymentGatewayResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/merchants/{id}": {
      "get": {
        "tags": ["Admin Merchant Management"],
        "summary": "Get merchant by ID",
        "description": "Returns a merchant by its database ID.",
        "operationId": "getMerchantById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Merchant retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperMerchantResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/gateway-channels/gateway/uid/{gatewayUid}": {
      "get": {
        "tags": ["Admin Gateway - Channel Management"],
        "summary": "Get channels for a gateway",
        "description": "Returns all channels linked to a given payment gateway in a paginated format.",
        "operationId": "getChannelsForGateway",
        "parameters": [
          {
            "name": "gatewayUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "Channels retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseWrapperPaymentGatewayChannelResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/v1/disbursements/uid/{uid}": {
      "get": {
        "tags": ["Admin Disbursement Management"],
        "summary": "Get disbursement by ID",
        "description": "Returns a disbursement by its database ID.",
        "operationId": "getDisbursementById",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Disbursement retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperDisbursementResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseWrapperUserResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/UserResponseDto" }
        }
      },
...
      "ResponseWrapperTransactionResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/TransactionResponseDto" }
        }
      },
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "uid": { "type": "string" },
          "internalTransactionId": { "type": "string" },
          "externalTransactionId": { "type": "string" },
          "merchantTransactionId": { "type": "string" },
          "pspTransactionId": { "type": "string" },
          "amount": { "type": "string" },
          "currency": { "type": "string" },
          "customerIdentifier": { "type": "string" },
          "paymentMethod": { "type": "string" },
          "customerName": { "type": "string" },
          "status": { "type": "string" },
          "colorCode": { "type": "string" },
          "errorCode": { "type": "string" },
          "errorMessage": { "type": "string" },
          "description": { "type": "string" },
          "pgoId": { "type": "string" },
          "pgoName": { "type": "string" },
          "merchantId": { "type": "string" },
          "merchantName": { "type": "string" },
          "submerchantId": { "type": "string" },
          "submerchantUid": { "type": "string" },
          "submerchantName": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        }
      },
...
      "ResponseWrapperPaymentGatewayResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/PaymentGatewayResponseDto" }
        }
      },
      "UpdatePaymentChannelRequestDto": {
        "type": "object",
        "description": "Request payload for updating an existing payment channel. The channel code is immutable and cannot be changed once created.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated descriptive name of the payment channel. Example: 'M-Pesa'",
            "maxLength": 50,
            "minLength": 0
          },
          "paymentChannelType": {
            "type": "string",
            "description": "Updated type of the payment channel. Accepted values: MNO, BANK, CARD, WALLET, etc.",
            "enum": ["MNO", "BANK", "CARD", "WALLET"]
          }
        }
      },
      "PaymentChannelResponseDto": {
        "type": "object",
        "description": "Represents a response payload for a payment channel entity.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Database ID of the payment channel",
            "example": 1
          },
          "uid": {
            "type": "string",
            "description": "Globally unique identifier (ULID) of the payment channel",
            "example": "01HZY4A7C1H0WXYZ123ABC4567"
          },
          "code": {
            "type": "string",
            "description": "Unique short code of the payment channel",
            "example": "MPESA"
          },
          "name": {
            "type": "string",
            "description": "Human-readable name of the payment channel",
            "example": "M-Pesa"
          },
          "paymentChannelType": {
            "type": "string",
            "description": "Type of payment channel (e.g., MNO, BANK, CARD)",
            "example": "MNO"
          },
          "isActive": {
            "type": "boolean",
            "description": "Flag indicating whether the channel is active and usable",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "description": "ISO-8601 timestamp when the payment channel was created",
            "example": "2025-08-04T14:33:21Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO-8601 timestamp when the payment channel was last updated",
            "example": "2025-08-04T14:34:01Z"
          }
        }
      },
      "ResponseWrapperPaymentChannelResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/PaymentChannelResponseDto" }
        }
      },
      "UpdateMerchantRequestDto": {
        "type": "object",
        "description": "Request object to update an existing merchant, including hierarchy and business details.",
        "properties": {
          "uid": {
            "type": "string",
            "description": "Merchant UID",
            "example": "01H8S9ABCDXYZ1234567WXYZ"
          },
          "merchantName": {
            "type": "string",
            "description": "Display name for the merchant",
            "example": "Safari Express Ltd",
            "maxLength": 100,
            "minLength": 0
          },
          "businessName": {
            "type": "string",
            "description": "Legal business name",
            "example": "Safari Express Company Ltd",
            "maxLength": 200,
            "minLength": 0
          },
          "businessRegistrationNumber": {
            "type": "string",
            "description": "Business registration number",
            "example": "BRN12345678",
            "maxLength": 50,
            "minLength": 0
          },
          "businessAddress": {
            "type": "string",
            "description": "Physical business address",
            "example": "Mikocheni A, Dar es Salaam",
            "maxLength": 200,
            "minLength": 0
          },
          "businessCity": {
            "type": "string",
            "description": "Business city",
            "example": "Dar es Salaam",
            "maxLength": 50,
            "minLength": 0
          },
          "businessState": {
            "type": "string",
            "description": "Business state or region",
            "example": "Kinondoni",
            "maxLength": 50,
            "minLength": 0
          },
          "businessPostalCode": {
            "type": "string",
            "description": "Postal code",
            "example": 14111,
            "maxLength": 20,
            "minLength": 0
          },
          "businessCountry": {
            "type": "string",
            "description": "Country of operation",
            "example": "Tanzania",
            "maxLength": 50,
            "minLength": 0
          },
          "contactEmail": {
            "type": "string",
            "format": "email",
            "description": "Merchant's primary email",
            "example": "support@safariexpress.co.tz",
            "maxLength": 100,
            "minLength": 0
          },
          "contactPhone": {
            "type": "string",
            "description": "Merchant's contact phone",
            "example": "+255712345678",
            "maxLength": 20,
            "minLength": 0
          },
          "websiteUrl": {
            "type": "string",
            "description": "Company website URL",
            "example": "https://www.safariexpress.co.tz",
            "maxLength": 100,
            "minLength": 0
          },
          "merchantType": {
            "type": "string",
            "description": "Type of merchant",
            "enum": [
              "Retail",
              "E-commerce",
              "Services",
              "Digital Goods",
              "Subscription",
              "Marketplace",
              "Financial Services",
              "Gaming",
              "Travel",
              "Food Delivery",
              "Other"
            ],
            "example": "TRAVEL"
          },
          "merchantRole": {
            "type": "string",
            "description": "Role of the merchant",
            "enum": ["Root", "Platform", "Sub-Merchant", "Agent", "Partner"],
            "example": "SUBMERCHANT"
          },
          "parentMerchantId": {
            "type": "integer",
            "format": "int64",
            "description": "Parent merchant ID (if applicable)",
            "example": 24
          },
          "singleTransactionLimit": {
            "type": "number",
            "description": "Optional limit for a single transaction",
            "example": 250000.0,
            "minimum": 0.0
          },
          "dailyTransactionLimit": {
            "type": "number",
            "description": "Optional limit for daily transaction volume",
            "example": 1000000.0,
            "minimum": 0.0
          },
          "monthlyTransactionLimit": {
            "type": "number",
            "description": "Optional limit for monthly transaction volume",
            "example": 5000000.0,
            "minimum": 0.0
          }
        },
        "required": ["merchantName"]
      },
...
      "ResponseWrapperMerchantResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/MerchantResponseDto" }
        }
      },
...
      "ResponseWrapperPaymentGatewayChannelResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": {
            "$ref": "#/components/schemas/PaymentGatewayChannelResponseDto"
          }
        }
      },
...
      "ResponseWrapperDisbursementResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/DisbursementResponseDto" }
        }
      },
...
      "ResponseWrapperBoolean": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "type": "boolean" }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "format": "int32", "minimum": 0 },
          "size": { "type": "integer", "format": "int32", "minimum": 1 },
          "sort": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "AssignChannelToGatewayRequestDto": {
        "type": "object",
        "description": "Request object for assigning a payment channel to a payment gateway",
        "properties": {
          "payCode": {
            "type": "string",
            "description": "Payment code used by the provider",
            "example": "MPESA_PAYBILL",
            "minLength": 1
          },
          "provider": {
            "type": "string",
            "description": "Provider name",
            "example": "Vodacom Tanzania",
            "minLength": 1
          }
        },
        "required": ["payCode", "provider"]
      },
...
      "PageResponseWrapperPaymentGatewayChannelResponseDto": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentGatewayChannelResponseDto"
            }
          },
          "pageNumber": { "type": "integer", "format": "int32" },
          "pageSize": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "last": { "type": "boolean" }
        }
      },
...
    },
    "securitySchemes": { "basicAuth": { "type": "http", "scheme": "basic" } }
  }
}